src/requirements.txt
pygame==2.1.2
numpy==1.24.1
black==22.12.0
pyyaml==6.0
colour==0.1.5